requirements.txt
googletrans==3.1.0a0
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
SQLAlchemy==1.4.36
vk_api==11.9.8
requests~=2.27.1
jsonlines~=3.0.0
langdetect~=1.0.9